requirements.txt
django>=4.2.6,<4.3.0
psycopg>=3.1.11,<3.2.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.3.0.<5.4.0
drf-spectacular>=0.26.5,<0.27.0
celery>=5.3.4,<5.4.0
redis>=5.0.0,<5.1.0
Pillow>=10.1.0,<10.2.0
